+++ plugins/js/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @hot-updater/js
 
+## 0.35.12
+
+### Patch Changes
+
+- 6e8b32e: Replace the semver dependency with verkit.
+  - @hot-updater/core@0.35.12
+
 ## 0.35.11
 
 ### Patch Changes